Implement FieldService.

+++ backend/Application/Interfaces/IFieldService.cs
@@ -4,7 +4,7 @@
     {
         internal Task<bool> SubjectContainsField(string subjectName, string fieldName, CancellationToken cancellationToken = default);
         internal Task<bool> SubjectExists(string subjectName, CancellationToken cancellationToken = default);
-        internal Task<(int SubjectKey, ICollection<int> FieldKeys)> GetSubjectAndFieldKeys(
+        internal Task<(long SubjectKey, ICollection<long> FieldKeys)> GetSubjectAndFieldKeys(
             string subjectName,
             ICollection<string> fieldName,
             CancellationToken cancellationToken = default);
